+++ omodsim/formatutils.h
@@ -20,8 +20,8 @@ inline QString formatUInt8Value(DataDisplayMode mode, quint8 c)
 {
     switch(mode)
     {
-        case DataDisplayMode::Decimal:
-        case DataDisplayMode::Integer:
+        case DataDisplayMode::UInt16:
+        case DataDisplayMode::Int16:
             return QString("%1").arg(QString::number(c), 3, '0');
 
         default:
@@ -41,8 +41,8 @@ inline QString formatUInt8Array(DataDisplayMode mode, const QByteArray& ar)
     for(quint8 i : ar)
         switch(mode)
         {
-            case DataDisplayMode::Decimal:
-            case DataDisplayMode::Integer:
+            case DataDisplayMode::UInt16:
+            case DataDisplayMode::Int16:
                 values += QString("%1").arg(QString::number(i), 3, '0');
             break;
 
@@ -69,8 +69,8 @@ inline QString formatUInt16Array(DataDisplayMode mode, const QByteArray& ar, Byt
         const quint16 value = makeUInt16(ar[i+1], ar[i], order);
         switch(mode)
         {
-            case DataDisplayMode::Decimal:
-            case DataDisplayMode::Integer:
+            case DataDisplayMode::UInt16:
+            case DataDisplayMode::Int16:
                 values += QString("%1").arg(QString::number(value), 5, '0');
                 break;
 
@@ -93,8 +93,8 @@ inline QString formatUInt16Value(DataDisplayMode mode, quint16 v)
 {
     switch(mode)
     {
-        case DataDisplayMode::Decimal:
-        case DataDisplayMode::Integer:
+        case DataDisplayMode::UInt16:
+        case DataDisplayMode::Int16:
             return QString("%1").arg(QString::number(v), 5, '0');
 
         default:
