+++ omodsim/modbusmultiserver.cpp
@@ -671,8 +671,8 @@ void ModbusMultiServer::writeRegister(QModbusDataUnit::RegisterType pointType, c
                 switch(params.DisplayMode)
                 {
                     case DataDisplayMode::Binary:
-                    case DataDisplayMode::Decimal:
-                    case DataDisplayMode::Integer:
+                    case DataDisplayMode::UInt16:
+                    case DataDisplayMode::Int16:
                     case DataDisplayMode::Hex:
                         data = createDataUnit(pointType, addr, params.Value.toUInt(), params.Order);
                     break;
